Huge Atmospherics Update - Five layer pipes and new funny gases (#10447)
* AHHHHHHHHHHHHH

* oopsie

* Fucking extools, who would of guessed

* eee

* MAPS GO BRRRRR

* Icons

* Update code/modules/atmospherics/machinery/portable/canister.dm


* Update code/modules/reagents/chemistry/reagents/other_reagents.dm


* Update code/modules/reagents/chemistry/reagents/other_reagents.dm


* Update code/modules/reagents/chemistry/reagents/other_reagents.dm


* Update code/game/objects/items/grenades/atmos_grenades.dm


* Update code/game/objects/items/twohanded.dm


* Update code/modules/atmospherics/machinery/components/binary_devices/pressure_valve.dm


* Update code/modules/atmospherics/gasmixtures/gas_mixture.dm


* Update code/modules/reagents/chemistry/reagents/other_reagents.dm


* Update code/modules/reagents/chemistry/reagents/other_reagents.dm


* More shit

* Update basemats.dm

* Cock and Balls

* oh my god kill me please I hate gases and colton is a slave master someone call the cops he is after me and I dont feel good. I need some help here because I dont know what im doing and im very scared.

* Fuck off linter

* HMMMMMMMM

* Hmm?

* afasggsag

* REEEE

* COMPI:E BAD

* Hmmmm

* icons maybe

* T mixer

* Icons

* Ree

* Please work

* Fucking yogs folder

* Reeee

* Icon and runtime

* Reee

* Indents

* MIXER GO BRRR

* E

* Last update of the night

* oopsie

* Linter go Brrr

* REEE

* MOre fixes

* e

* Electroshitter

* EEE

* FUCK

* Maybe makes this work

* tg bad

* Come at me linter.

* Fucking oops

* Update electrolyzer.dm

* LINTER GO TO HELL - GIMMIE DAT GREEN TICK

* Im dumb

* E

* Update electrolyzer.dm

* Fixes vents

* Fixes

* E

* Fucking Jank code is jank

* R

* r

* Update tgui.bundle.js

* fucking moja just used goddamn find and replace

* Theos

+++ code/modules/atmospherics/environmental/LINDA_fire.dm
@@ -18,16 +18,18 @@
 		return
 	var/tox = air.get_moles(/datum/gas/plasma)
 	var/trit = air.get_moles(/datum/gas/tritium)
+	var/h2 = air.get_moles(/datum/gas/hydrogen)
+
 	if(active_hotspot)
 		if(soh)
-			if(tox > 0.5 || trit > 0.5)
+			if(tox > 0.5 || trit > 0.5 || h2 > 0.5)
 				if(active_hotspot.temperature < exposed_temperature)
 					active_hotspot.temperature = exposed_temperature
 				if(active_hotspot.volume < exposed_volume)
 					active_hotspot.volume = exposed_volume
 		return
 
-	if((exposed_temperature > PLASMA_MINIMUM_BURN_TEMPERATURE) && (tox > 0.5 || trit > 0.5))
+	if((exposed_temperature > PLASMA_MINIMUM_BURN_TEMPERATURE) && (tox > 0.5 || trit > 0.5 || h2 > 0.5))
 
 		active_hotspot = new /obj/effect/hotspot(src, exposed_volume*25, exposed_temperature)
 
@@ -166,7 +168,7 @@
 		return
 
 	//Not enough / nothing to burn
-	if(!location.air || (INSUFFICIENT(/datum/gas/plasma) && INSUFFICIENT(/datum/gas/tritium)) || INSUFFICIENT(/datum/gas/oxygen))
+	if(!location.air || (INSUFFICIENT(/datum/gas/plasma) && INSUFFICIENT(/datum/gas/tritium)) && INSUFFICIENT(/datum/gas/hydrogen) || INSUFFICIENT(/datum/gas/oxygen))
 		qdel(src)
 		return
